user-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis==5.0.1
aioredis==2.0.1
celery==5.3.4
pika==1.3.2
aio-pika==9.3.1
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
PyJWT==2.8.0
cryptography==41.0.7
python-multipart==0.0.6
email-validator==2.1.0
jinja2==3.1.2
python-dotenv==1.0.0
pymongo==4.6.0
motor==3.3.2
geopy==2.4.1
shapely==2.0.2
geoalchemy2==0.14.2
websockets==12.0
socketio<5.13.0
python-socketio==5.10.0
aiofiles==23.2.1
pillow==10.1.0
reportlab==4.0.7
weasyprint==60.2
requests==2.31.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
factory-boy==3.3.0
faker==20.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pyotp==2.8.0
qrcode[pil]==7.4.2

